chore: move App.tsx
